+++ linkerd/core/src/main/scala/io/buoyant/linkerd/Server.scala
@@ -4,15 +4,12 @@ import com.fasterxml.jackson.annotation.JsonIgnore
 import com.github.ghik.silencer.silent
 import com.twitter.concurrent.AsyncSemaphore
 import com.twitter.conversions.time._
+import com.twitter.finagle.buoyant.TlsServerConfig
 import com.twitter.finagle.filter.RequestSemaphoreFilter
-import com.twitter.finagle.ssl._
-import com.twitter.finagle.ssl.server.{LegacyKeyServerEngineFactory, SslServerConfiguration, SslServerEngineFactory}
-import com.twitter.finagle.transport.Transport
 import com.twitter.finagle.service.TimeoutFilter
-import com.twitter.finagle.ssl.client.SslClientEngineFactory
+import com.twitter.finagle.ssl.server.{LegacyKeyServerEngineFactory, SslServerEngineFactory}
 import com.twitter.finagle.{ListeningServer, Path, Stack}
 import io.buoyant.config.types.Port
-import java.io.File
 import java.net.{InetAddress, InetSocketAddress}
 
 /**
@@ -40,12 +37,6 @@ trait Server {
 }
 
 object Server {
-
-  case class RouterLabel(label: String)
-  implicit object RouterLabel extends Stack.Param[RouterLabel] {
-    val default = RouterLabel("")
-  }
-
   /**
    * A [[Server]] that is fully configured but not yet listening.
    */
@@ -80,7 +71,6 @@ object Server {
 
     override def withParams(ps: Stack.Params): Server = copy(params = ps)
   }
-
 }
 
 class ServerConfig { config =>
@@ -100,34 +90,10 @@ class ServerConfig { config =>
 
   @JsonIgnore
   protected def serverParams: Stack.Params = Stack.Params.empty
-    .maybeWith(tls.map(tls(_)))
+    .maybeWith(tls.map(_.params(alpnProtocols, sslServerEngine)))
     .maybeWith(clearContext.map(ClearContext.Enabled(_)))
     .maybeWith(timeoutMs.map(timeout => TimeoutFilter.Param(timeout.millis))) +
     RequestSemaphoreFilter.Param(requestSemaphore)
-
-  @JsonIgnore
-  private[this] def tls(c: TlsServerConfig): Stack.Params = {
-    assert(c.certPath != null)
-    assert(c.keyPath != null)
-    val trust = c.caCertPath match {
-      case Some(caCertPath) => TrustCredentials.CertCollection(new File(caCertPath))
-      case None => TrustCredentials.Unspecified
-    }
-    val ciphers = c.ciphers match {
-      case Some(cs) => CipherSuites.Enabled(cs)
-      case None => CipherSuites.Unspecified
-    }
-    val appProtocols = alpnProtocols match {
-      case Some(ps) => ApplicationProtocols.Supported(ps)
-      case None => ApplicationProtocols.Unspecified
-    }
-    Stack.Params.empty + Transport.ServerSsl(Some(SslServerConfiguration(
-      keyCredentials = KeyCredentials.CertAndKey(new File(c.certPath), new File(c.keyPath)),
-      trustCredentials = trust,
-      cipherSuites = ciphers,
-      applicationProtocols = appProtocols
-    ))) + SslServerEngineFactory.Param(sslServerEngine)
-  }
 
   @JsonIgnore
   def alpnProtocols: Option[Seq[String]] = None
@@ -150,10 +116,3 @@ class ServerConfig { config =>
     announce.toSeq.flatten.map(Path.read)
   )
 }
-
-case class TlsServerConfig(
-  certPath: String,
-  keyPath: String,
-  caCertPath: Option[String],
-  ciphers: Option[Seq[String]]
-)