+++ linkerd/protocol/thrift/src/test/scala/io/buoyant/linkerd/protocol/ThriftInitializerTest.scala
@@ -3,7 +3,6 @@ package io.buoyant.linkerd.protocol
 import com.fasterxml.jackson.databind.JsonMappingException
 import com.twitter.finagle.Path
 import com.twitter.finagle.Thrift.param
-import com.twitter.finagle.Thrift.param.AttemptTTwitterUpgrade
 import io.buoyant.linkerd.Linker
 import io.buoyant.router.StackRouter.Client.PerClientParams
 import io.buoyant.router.Thrift.param.MethodInDst
@@ -17,33 +16,33 @@ class ThriftInitializerTest extends FunSuite with Exceptions {
     val config = """
       |routers:
       |- protocol: thrift
+      |  thriftProtocol: compact
       |  thriftMethodInDst: true
       |  client:
       |    thriftFramed: false
-      |    thriftProtocol: binary
       |    attemptTTwitterUpgrade: false
       |  servers:
       |  - thriftFramed: true
-      |    thriftProtocol: compact
     """.stripMargin
 
     val linker = Linker.Initializers(Seq(ThriftInitializer)).load(config)
     val router = linker.routers.head
-    val params = router.params[PerClientParams].paramsFor(Path.read("/foo"))
-    assert(router.params[MethodInDst].enabled)
-    assert(!params[param.Framed].enabled)
-    assert(!params[param.ProtocolFactory].protocolFactory.isInstanceOf[TCompactProtocol.Factory])
-    assert(router.servers.head.params[param.Framed].enabled)
-    assert(router.servers.head.params[param.ProtocolFactory].protocolFactory.isInstanceOf[TCompactProtocol.Factory])
-    assert(!params[AttemptTTwitterUpgrade].upgrade)
+    val routerParams = router.params
+    val serverParams = router.servers.head.params
+    val clientParams = router.params[PerClientParams].paramsFor(Path.read("/foo"))
+    assert(routerParams[MethodInDst].enabled)
+    assert(routerParams[param.ProtocolFactory].protocolFactory.isInstanceOf[TCompactProtocol.Factory])
+    assert(serverParams[param.Framed].enabled)
+    assert(serverParams[param.ProtocolFactory].protocolFactory.isInstanceOf[TCompactProtocol.Factory])
+    assert(!clientParams[param.Framed].enabled)
+    assert(!clientParams[param.AttemptTTwitterUpgrade].upgrade)
   }
 
   test("unsupported thrift protocol") {
     val config = """
       |routers:
       |- protocol: thrift
-      |  servers:
-      |    thriftProtocol: magic
+      |  thriftProtocol: magic
     """.stripMargin
 
     assertThrows[JsonMappingException] {
@@ -55,13 +54,13 @@ class ThriftInitializerTest extends FunSuite with Exceptions {
     val config = """
                    |routers:
                    |- protocol: thrift
+                   |  thriftProtocol: compact
                    |  servers:
                    |  - thriftFramed: true
-                   |    thriftProtocol: compact
                  """.stripMargin
 
     val linker = Linker.Initializers(Seq(ThriftInitializer)).load(config)
     val params = linker.routers.head.params[PerClientParams].paramsFor(Path.read("/foo"))
-    assert(!params[AttemptTTwitterUpgrade].upgrade)
+    assert(!params[param.AttemptTTwitterUpgrade].upgrade)
   }
 }