+++ namerd/iface/control-http/src/main/scala/io/buoyant/namerd/iface/HttpControlServiceConfig.scala
@@ -3,7 +3,8 @@ package io.buoyant.namerd.iface
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.twitter.finagle.naming.NameInterpreter
 import com.twitter.finagle.stats.StatsReceiver
-import com.twitter.finagle.{Namer, Path, Http, ListeningServer}
+import com.twitter.finagle.{Http, ListeningServer, Namer, Path, Stack}
+import com.twitter.finagle.param
 import io.buoyant.namerd._
 import java.net.{InetAddress, InetSocketAddress}
 
@@ -13,7 +14,20 @@ class HttpControlServiceConfig extends InterpreterInterfaceConfig {
     namers: Map[Path, Namer],
     store: DtabStore,
     stats: StatsReceiver
+<<<<<<< HEAD
   ): Servable = HttpControlServable(addr, store, delegate, namers, stats.scope(HttpControlServiceConfig.kind))
+=======
+  ): Servable = {
+    val iface = new HttpControlService(store, delegate, namers)
+    val params =
+      tlsParams +
+        param.Stats(stats) +
+        param.Label(HttpControlServiceConfig.kind) +
+        Http.Netty4Impl
+
+    HttpControlServable(addr, iface, params)
+  }
+>>>>>>> master
 
   @JsonIgnore
   def defaultAddr = HttpControlServiceConfig.defaultAddr
@@ -26,16 +40,19 @@ object HttpControlServiceConfig {
 
 case class HttpControlServable(
   addr: InetSocketAddress,
-  store: DtabStore,
-  delegate: Ns => NameInterpreter,
-  namers: Map[Path, Namer],
-  stats: StatsReceiver
+  iface: HttpControlService,
+  params: Stack.Params
 ) extends Servable {
   def kind = HttpControlServiceConfig.kind
+<<<<<<< HEAD
   def serve(): ListeningServer = Http.server
+=======
+  val http = Http.server
+  def serve(): ListeningServer = http
+    .withParams(http.params ++ params)
+>>>>>>> master
     .withStreaming(true)
-    .withStatsReceiver(stats)
-    .serve(addr, new HttpControlService(store, delegate, namers))
+    .serve(addr, iface)
 }
 
 class HttpControlServiceInitializer extends InterfaceInitializer {