+++ namerd/iface/interpreter-thrift/src/main/scala/io/buoyant/namerd/iface/ThriftInterpreterInterfaceConfig.scala
@@ -1,12 +1,8 @@
 package io.buoyant.namerd.iface
 
 import com.fasterxml.jackson.annotation.JsonIgnore
-import com.github.ghik.silencer.silent
-import com.twitter.finagle.ssl.{KeyCredentials, Ssl}
-import com.twitter.finagle.ssl.server.{LegacyKeyServerEngineFactory, SslServerConfiguration, SslServerEngineFactory}
 import com.twitter.finagle.stats.StatsReceiver
-import com.twitter.finagle.transport.Transport
-import com.twitter.finagle.{Namer, Path, Stack, ThriftMux}
+import com.twitter.finagle.{Namer, Path, Stack, Thrift, ThriftMux}
 import com.twitter.finagle.naming.NameInterpreter
 import com.twitter.finagle.param
 import com.twitter.scrooge.ThriftService
@@ -14,15 +10,13 @@ import com.twitter.util.Duration
 import com.twitter.util.TimeConversions._
 import io.buoyant.namerd.iface.ThriftNamerInterface.LocalStamper
 import io.buoyant.namerd._
-import java.io.File
 import java.net.InetSocketAddress
 import scala.util.Random
 
 case class ThriftInterpreterInterfaceConfig(
   retryBaseSecs: Option[Int] = None,
   retryJitterSecs: Option[Int] = None,
-  cache: Option[CapacityConfig] = None,
-  tls: Option[TlsServerConfig] = None
+  cache: Option[CapacityConfig] = None
 ) extends InterpreterInterfaceConfig {
   @JsonIgnore
   protected def defaultAddr = ThriftInterpreterInterfaceConfig.defaultAddr
@@ -49,10 +43,10 @@ case class ThriftInterpreterInterfaceConfig(
       cache.map(_.capacity).getOrElse(ThriftNamerInterface.Capacity.default),
       stats1
     )
-    val params = (tls match {
-      case Some(tlsConfig) => tlsConfig.params
-      case None => Stack.Params.empty
-    }) + param.Stats(stats1)
+    val params =
+      tlsParams +
+        param.Stats(stats1) +
+        Thrift.ThriftImpl.Netty4
     ThriftServable(addr, iface, params)
   }
 }
@@ -87,14 +81,4 @@ case class CapacityConfig(
     addrCacheActive = addrCacheActive.getOrElse(default.addrCacheActive),
     addrCacheInactive = addrCacheInactive.getOrElse(default.addrCacheInactive)
   )
-}
-
-case class TlsServerConfig(certPath: String, keyPath: String) {
-  val params = {
-    val creds = KeyCredentials.CertAndKey(new File(certPath), new File(keyPath))
-    @silent val sslServerEngine = SslServerEngineFactory.Param(LegacyKeyServerEngineFactory)
-    Stack.Params.empty +
-      Transport.ServerSsl(Some(SslServerConfiguration(keyCredentials = creds))) +
-      sslServerEngine
-  }
 }